Fix new rubocop Style/OneClassPerFile warnings
