+++ .rubocop_todo.yml
@@ -179,19 +179,6 @@ Style/IfUnlessModifier:
 Style/NumericLiterals:
   MinDigits: 15
 
-# Offense count: 17
-# Configuration parameters: AllowedClasses.
-Style/OneClassPerFile:
-  Exclude:
-    - 'config/initializers/i18n.rb'
-    - 'db/migrate/20120208122334_merge_acl_address_and_mask.rb'
-    - 'test/abilities/abilities_test.rb'
-    - 'test/abilities/api_abilities_test.rb'
-    - 'test/abilities/api_capability_test.rb'
-    - 'test/system/embed_test.rb'
-    - 'test/validators/characters_validator_test.rb'
-    - 'test/validators/whitespace_validator_test.rb'
-
 # Offense count: 705
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: AllowHeredoc, AllowURI, AllowQualifiedName, URISchemes, AllowRBSInlineAnnotation, AllowCopDirectives, AllowedPatterns, SplitStrings.
